requirements.txt
anyio==4.2.0
certifi==2023.11.17
flake8==6.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.1.0
python-dotenv==1.0.0
sniffio==1.3.0
